Implement app players methods

+++ internal/model/vote.go
@@ -5,5 +5,5 @@ import "github.com/oklog/ulid/v2"
 type VoteRecord struct {
 	UserID       string
 	PollPlayerID ulid.ULID `json:"pollPlayerId"`
-	AnswerIDs    []int     `json:"answerIds"`
+	AnswerIDs    []int64   `json:"answerIds"`
 }
